Create a new file for the extension NotificationCenter
